requirements.txt
streamlit
pymongo[srv]
python-dotenv
google-generativeai
requests
numpy
pandas
regex
